GameEnd 수정
타임아웃 시 무승부인 팀들이 있을 수 있어서 이긴 팀을 비트플래그로 전송

+++ Client/Assets/02. Scripts/Network/PacketProcessor/Event/BombSpawn.cs
@@ -18,7 +18,7 @@ namespace Network.PacketProcessor
             int posY = Data.Pos.Value.Y;
             int id = Data.Id;
 
-            Debug.Log("Spawn Bomb : " + id);
+            Debug.Log("Spawn Bomb : " + (int)id);
             // 폭탄 생성
             Managers.BombObject.SpawnBomb(posX, posY, id);
         }
